requirements.txt
fabric==2.4.0
paho_mqtt==1.4.0
yeelight==0.4.3
